+++ mtg_proxy_printer/ui/main_window.py
@@ -16,7 +16,8 @@
 import pathlib
 import typing
 
-from PyQt5.QtCore import pyqtSlot, pyqtSignal, QStringListModel, QItemSelectionModel
+
+from PyQt5.QtCore import pyqtSlot, pyqtSignal, QStringListModel
 from PyQt5.QtGui import QCloseEvent, QResizeEvent, QShowEvent, QKeySequence
 from PyQt5.QtWidgets import QApplication, QMessageBox, QProgressBar, QAction, QWidget, QToolBar
 
@@ -27,9 +28,7 @@ from mtg_proxy_printer.model.document import Document
 import mtg_proxy_printer.settings
 import mtg_proxy_printer.print
 from mtg_proxy_printer.ui.common import inherits_from_ui_file_with_name
-from mtg_proxy_printer.ui.current_page_view import CurrentPageView
-from mtg_proxy_printer.ui.document_view import DocumentView
-from mtg_proxy_printer.ui.add_card import AddCardWidget
+from mtg_proxy_printer.ui.central_widget import CentralWidgetTypes, get_configured_central_widget_layout_class
 from mtg_proxy_printer.ui.dialogs import SavePDFDialog, SaveDocumentAsDialog, LoadDocumentDialog, \
     AboutMTGProxyPrinterDialog, PrintPreviewDialog, PrintDialog, DocumentSettingsDialog
 from mtg_proxy_printer.ui.cache_cleanup_wizard import CacheCleanupWizard
@@ -38,13 +37,12 @@ from mtg_proxy_printer.ui.deck_import_wizard import DeckImportWizard
 from mtg_proxy_printer.logger import get_logger
 logger = get_logger(__name__)
 del get_logger
-layout = mtg_proxy_printer.settings.settings["gui"]["search-widget-layout"]
 __all__ = [
     "MainWindow",
 ]
 
 
-class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_window")):
+class MainWindow(*inherits_from_ui_file_with_name(f"main_window")):
 
     should_update_languages = pyqtSignal()
     window_size_changed = pyqtSignal()
@@ -59,7 +57,7 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
                  language_model: QStringListModel,
                  *args, **kwargs):
         super(MainWindow, self).__init__(*args, **kwargs)
-        logger.info(f"Creating {self.__class__.__name__} instance using the {layout} layout.")
+        logger.info(f"Creating {self.__class__.__name__} instance.")
         self.card_data_download_in_progress = False
         self.setupUi(self)
         self.about_dialog = self._create_about_dialog()
@@ -72,20 +70,15 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
         self.language_model = language_model
         self.card_data_downloader = card_info_downloader
         self._connect_card_info_downloader_signals(card_info_downloader)
-        self.page_view: CurrentPageView
-        self._setup_page_view(document)
+        self.central_widget: CentralWidgetTypes
+        self._setup_central_widget()
         self._setup_loading_state_connections()
-        self._setup_add_card_widget(card_db, image_db)
-        self._setup_document_view(document)
-        self.action_new_page.triggered.connect(document.add_page)
         self.should_update_languages.connect(
             lambda: self.language_model.setStringList(self.card_database.get_all_languages())
         )
-        self.should_update_languages.connect(self.add_card_widget.update_selected_language)
-        self.settings_changed.connect(self.add_card_widget.update_selected_language)
+        self.should_update_languages.connect(self.central_widget.add_card_widget.update_selected_language)
         self.settings_changed.connect(document.apply_settings)
-        self.settings_changed.connect(self.page_view.settings_changed)
-        document.page_layout_changed.connect(self.page_view.settings_changed)
+        self.settings_changed.connect(self.central_widget.settings_changed)
         self.settings_changed.connect(self.offer_re_downloading_card_database)
         self.action_show_toolbar: QAction
         self.action_show_toolbar.setChecked(mtg_proxy_printer.settings.settings["gui"].getboolean("show-toolbar"))
@@ -111,11 +104,15 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
         for action, shortcut in actions_with_shortcuts:
             action.setShortcut(shortcut)
 
-    def _setup_page_view(self, document: Document):
-        self.page_view: CurrentPageView
-        self.page_view.set_document(document)
-        self.window_size_changed.connect(self.page_view.window_size_changed)
-        self.document.current_page_changed.connect(self.page_view.on_current_page_changed)
+    def _setup_central_widget(self):
+        central_widget_class = get_configured_central_widget_layout_class()
+        logger.debug(f"Using central widget class {central_widget_class.__name__}")
+        self.central_widget: CentralWidgetTypes
+        self.central_widget = central_widget_class(self)
+        self.setCentralWidget(self.central_widget)
+        self.central_widget.set_data(self.document, self.card_database, self.image_db)
+        self.action_discard_page.triggered.connect(self.central_widget.on_action_discard_page_triggered)
+        self.window_size_changed.connect(self.central_widget.window_size_changed)
 
     def _setup_loading_state_connections(self):
         for widget_or_action in self._get_widgets_and_actions_disabled_in_loading_state():
@@ -126,7 +123,7 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
         document.loader.loading_file_failed.connect(self.on_document_loading_failed)
         document.loader.unknown_scryfall_ids_found.connect(self.on_document_loading_found_unknown_scryfall_ids)
         document.loader.network_error_occurred.connect(self.on_network_error_occurred)
-        document.loading_state_changed.connect(self._select_first_page)
+        self.action_new_page.triggered.connect(document.add_page)
         self.action_new_document.triggered.connect(document.clear_all_data)
         self.action_compact_document.triggered.connect(document.compact_pages)
 
@@ -160,10 +157,9 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
             self.action_import_deck_list,
             self.action_new_page,
             self.action_discard_page,
-            self.add_card_widget,
             self.action_show_settings,
             self.action_cleanup_local_image_cache,
-            self.page_view.delete_selected_images_button,
+            self.central_widget,
         ]
 
     def _connect_image_database_signals(self, image_db: ImageDatabase):
@@ -179,17 +175,6 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
         self.statusBar().addPermanentWidget(progress_bar)
         return progress_bar
 
-    def _setup_add_card_widget(self, card_db: CardDatabase, image_db: ImageDatabase):
-        self.add_card_widget: AddCardWidget
-        self.add_card_widget.set_card_database(card_db)
-        self.add_card_widget.card_added.connect(image_db.get_new_card_image_asynchronous)
-
-    def _setup_document_view(self, document: Document):
-        self.document_view: DocumentView
-        self.document_view.setModel(document)
-        self.document_view.selectionModel().currentChanged.connect(document.on_ui_selects_new_page)
-        self._select_first_page()
-
     def offer_re_downloading_card_database(self):
         settings_changed = self.card_database.check_if_download_settings_changed()
         self.action_download_card_data.setEnabled(self.card_database.allow_updating_card_data())
@@ -201,14 +186,6 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
                 QMessageBox.Yes | QMessageBox.No
                 ) == QMessageBox.Yes:
             self.action_download_card_data.trigger()
-
-    @pyqtSlot()
-    def _select_first_page(self, loading_in_progress: bool = False):
-        if not loading_in_progress:
-            logger.info("Loading finished. Selecting first page.")
-            new_selection = self.document.index(0, 0)
-            self.document_view.selectionModel().select(new_selection, QItemSelectionModel.Select)
-            self.document.on_ui_selects_new_page(new_selection)
 
     def resizeEvent(self, event: QResizeEvent):
         super(MainWindow, self).resizeEvent(event)
@@ -346,32 +323,6 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
         self.progress_bar.show()
 
     @pyqtSlot()
-    def on_action_discard_page_triggered(self):
-        self.document_view: DocumentView
-        if self.document.rowCount() == 1:
-            logger.info(f"User selects to delete the only page, so clearing it.")
-            self.document.clear_page(self.document.index(0, 0))
-            return
-        to_be_deleted: int = self.document_view.selectedIndexes()[0].row()
-        logger.info(f"User selects to delete the currently selected page. Will be removing page {to_be_deleted}")
-        logger.debug("Deleting the requested page.")
-        # TODO: Investigate, why unsetting the model is needed.
-        #  The document_view’s selection model somehow asks for data using invalid
-        #  indices, when the last page is selected and gets deleted. The only way around seems to be to
-        #  completely disconnect the model, remove the row, then set it again.
-        self.document_view.setModel(None)
-        self.document.remove_pages([self.document.index(to_be_deleted, 0)])
-        # Now reset the model (and reconnect the currentChanged signal, which seems to be disconnected implicitly
-        self.document_view.setModel(self.document)
-        self.document_view.selectionModel().currentChanged.connect(self.document.on_ui_selects_new_page)
-
-        new_row_index = min(to_be_deleted, self.document.rowCount() - 1)
-        logger.debug(f"Selecting page {new_row_index}.")
-        new_row_selection = self.document.index(new_row_index, 0)
-        self.document_view.selectionModel().select(new_row_selection, QItemSelectionModel.Select)
-        self.document.on_ui_selects_new_page(new_row_selection)
-
-    @pyqtSlot()
     def on_action_save_document_triggered(self):
         logger.debug("User clicked on Save")
         if self.document.save_file_path is None:
@@ -397,7 +348,7 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
     def on_action_load_document_triggered(self):
         dialog = LoadDocumentDialog(self, self.document)
         if dialog.exec_() == LoadDocumentDialog.Accepted:
-            self._select_first_page()
+            self.central_widget.select_first_page()
 
     def on_document_loading_failed(self, failed_path: pathlib.Path, reason: str):
         QMessageBox.critical(
@@ -437,27 +388,27 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
             # If the user declines to perform the update now, allow them to perform it later by enabling the action.
             self.action_download_card_data.setEnabled(True)
 
-    def ask_user_about_application_update_policy(self) -> bool:
+    def ask_user_about_application_update_policy(self):
         """Executed on start when the application update policy setting is set to None, the default value."""
         name = mtg_proxy_printer.meta_data.PROGRAMNAME
-        return self._ask_user_about_update_policy(
+        self._ask_user_about_update_policy(
             title="Check for application updates?",
             question=f"Automatically check for application updates whenever you start {name}?",
             logger_message="Application update policy set.",
             settings_key="check-for-application-updates"
         )
 
-    def ask_user_about_card_data_update_policy(self) -> bool:
+    def ask_user_about_card_data_update_policy(self):
         """Executed on start when the card data update policy setting is set to None, the default value."""
         name = mtg_proxy_printer.meta_data.PROGRAMNAME
-        return self._ask_user_about_update_policy(
+        self._ask_user_about_update_policy(
             title="Check for card data updates?",
             question=f"Automatically check for card data updates on Scryfall whenever you start {name}?",
             logger_message="Card data update policy set.",
             settings_key="check-for-card-data-updates"
         )
 
-    def _ask_user_about_update_policy(self, title: str, question: str, logger_message: str, settings_key: str) -> bool:
+    def _ask_user_about_update_policy(self, title: str, question: str, logger_message: str, settings_key: str):
         if (result := QMessageBox.question(
                 self, title,
                 f"{question}\nYou can change this later in the settings.",
@@ -468,5 +419,3 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
                 result == QMessageBox.Yes)
             mtg_proxy_printer.settings.write_settings_to_file()
             logger.debug("Written settings to disk.")
-            return True
-        return False